Player now takes damage whilst key systems are being destroyed

+++ core/src/com/mygdx/auber/Scenes/Hud.java
@@ -40,7 +40,7 @@ public class Hud {
 
         imposterCountLabel = new Label(String.format("Imposter Arrests: %02d / %02d", ImposterCount, PlayScreen.numberOfInfiltrators), new Label.LabelStyle(new BitmapFont(), Color.GREEN));
         crewmateCountLabel = new Label(String.format("Incorrect crewmate Arrests: %02d / %02d", CrewmateCount, PlayScreen.maxIncorrectArrests), new Label.LabelStyle(new BitmapFont(), Color.RED));
-        playerHealthLabel = new Label(String.format("Health: %02d", Player.health), new Label.LabelStyle(new BitmapFont(), Color.RED));
+        playerHealthLabel = new Label(String.format("Health: %02d", (int) Player.health), new Label.LabelStyle(new BitmapFont(), Color.RED));
 
         hudTable.add(imposterCountLabel).expandX().left().padLeft(10);
         hudTable.add(crewmateCountLabel).expandX().right().padRight(10);
@@ -54,7 +54,7 @@ public class Hud {
     public void update() {
         imposterCountLabel.setText(String.format("Imposter Arrests: %02d / %02d", ImposterCount, PlayScreen.numberOfInfiltrators));
         crewmateCountLabel.setText(String.format("Incorrect crewmate Arrests: %02d / %02d", CrewmateCount, PlayScreen.maxIncorrectArrests));
-        playerHealthLabel.setText(String.format("Health: %02d", Player.health));
+        playerHealthLabel.setText(String.format("Health: %02d",(int) Player.health));
     }
 
 }
